- switched to chgres_cube program to create FV3ICs and add aerosol variables  -- checkout feature/gdas_init.v16 branch from George Gayno's UFS_UTILS fork - updated sorc/fv3gfs.fd to 18mar gsl/develop branch, f204bfd - updated namelist parameters for GSL physics    k_split=6    n_split=2 - added dz_min=2, psm_bc=1 to exglobal forecast script - set convert_dry_rho to .true. in sorc/fv3gfs.fd/FV3/ccpp/GFS_typedefs.F90 - added init and updated getic tasks for v16 initial conditions - updated archive script

+++ jobs/rocoto/init.sh
@@ -29,7 +29,7 @@ done
 
 ###############################################################
 # Source machine runtime environment
-. $BASE_ENV/${machine}.env getic
+. $BASE_ENV/${machine}.env init
 status=$?
 [[ $status -ne 0 ]] && exit $status
 
@@ -45,69 +45,39 @@ export DATA=${DATA:-${DATAROOT}/init}
 export EXTRACT_DIR=${EXTRACT_DIR:-$ROTDIR}
 export WORKDIR=${WORKDIR:-$DATA}
 export OUTDIR=${OUTDIR:-$ROTDIR}
-export PRODHPSSDIR=${PRODHPSSDIR:-/NCEPPROD/hpssprod/runhistory}
 export COMPONENT="atmos"
 export gfs_ver=${gfs_ver:-v16}
-export GETICSH=${GETICSH:-${GDASINIT_DIR}/get_v16.data.sh}
-
-# Run get data script
-if [ ! -d $EXTRACT_DIR ]; then mkdir -p $EXTRACT_DIR ; fi
+export RUNICSH=${RUNICSH:-${GDASINIT_DIR}/run_v16.chgres.sh}
 
 # Check if init is needed and run if so
 if [[ $gfs_ver = "v16" && $EXP_WARM_START = ".true." && $CASE = "C768" ]]; then
-  # Pull RESTART files off HPSS
-
-  if [ ! -d $ROTDIR ]; then mkdir $ROTDIR ; fi
-  cd $ROTDIR
-
-  if [ ${RETRO:-"NO"} = "YES" ]; then # Retrospective parallel input
-
-    # Pull prior cycle restart files
-    BDATE=`$NDATE -06 ${yy}${mm}${dd}${hh}`
-    htar -xvf ${HPSSDIR}/${BDATE}/gdas_restartb.tar
-    status=$?
-    [[ $status -ne 0 ]] && exit $status
-
-    if [ $DO_WAVE = "YES" ]; then
-      # Pull prior cycle wave restart files
-      htar -xvf ${HPSSDIR}/${BDATE}/gdaswave_restart.tar
-      status=$?
-      [[ $status -ne 0 ]] && exit $status
-    fi
-
-    # Pull current cycle restart files
-    htar -xvf ${HPSSDIR}/${CDATE}/gfs_restarta.tar
-    status=$?
-    [[ $status -ne 0 ]] && exit $status
-
-    # Pull IAU increment files
-    htar -xvf ${HPSSDIR}/${CDATE}/gfs_netcdfa.tar
-    status=$?
-    [[ $status -ne 0 ]] && exit $status
-
-# else # Opertional input
-#   # ADD AFTER IMPLEMENTATION
-  fi
-
+  echo "Detected v16 C768 warm starts, will not run init. Exiting..."
+  exit 0
 else
-
-  # Run UFS_UTILS GETICSH
-  sh ${GETICSH} ${CDUMP}
+  # Run chgres_cube
+  if [ ! -d $OUTDIR ]; then mkdir -p $OUTDIR ; fi
+  sh ${RUNICSH} ${CDUMP}
   status=$?
   [[ $status -ne 0 ]] && exit $status
-
 fi
 
-# Pull pgbanl file for verification/archival - v14+
+# Check for pgbanl files and move
 if [ $gfs_ver = v14 -o $gfs_ver = v15 ]; then
   cd $EXTRACT_DIR
   for grid in 0p25 0p50 1p00
   do
     file=gfs.t${hh}z.pgrb2.${grid}.anl
-    htar -xvf ${PRODHPSSDIR}/rh${yy}/${yy}${mm}/${yy}${mm}${dd}/${tarball} ./${CDUMP}.${yy}${mm}${dd}/${hh}/${file}
-    mv ${EXTRACT_DIR}/${CDUMP}.${yy}${mm}${dd}/${hh}/${file} ${OUTDIR}/${CDUMP}.${yy}${mm}${dd}/${hh}/${COMPONENT}/${file}
+    if [ -f ${EXTRACT_DIR}/${CDUMP}.${yy}${mm}${dd}/${hh}/${file} ]; then
+      mv ${EXTRACT_DIR}/${CDUMP}.${yy}${mm}${dd}/${hh}/${file} ${OUTDIR}/${CDUMP}.${yy}${mm}${dd}/${hh}/${COMPONENT}/${file}
+    fi
   done
 fi
+
+##########################################
+# Remove the Temporary working directory
+##########################################
+cd $DATAROOT
+[[ $KEEPDATA = "NO" ]] && rm -rf $DATA
 
 ###############################################################
 # Exit out cleanly
